+++ cmd/create/flag.go
@@ -23,7 +23,6 @@ const (
 	flagRelease          = "release"
 	flagReleases         = "releases"
 	flagToken            = "token"
-	flagWait             = "wait"
 )
 
 type flag struct {
@@ -36,7 +35,6 @@ type flag struct {
 	Release          string
 	Releases         string
 	Token            string
-	Wait             bool
 }
 
 func (f *flag) Init(cmd *cobra.Command) {
@@ -49,7 +47,6 @@ func (f *flag) Init(cmd *cobra.Command) {
 	cmd.Flags().StringVarP(&f.Release, flagRelease, "r", "", fmt.Sprintf(`The semantic version of the release to be tested.`))
 	cmd.Flags().StringVarP(&f.Releases, flagReleases, "s", "", fmt.Sprintf(`The path of the releases repo on the local filesystem.`))
 	cmd.Flags().StringVarP(&f.Token, flagToken, "t", "", `The token used to authenticate with the GS API.`)
-	cmd.Flags().BoolVarP(&f.Wait, flagWait, "w", false, `If true, wait for the test cluster API to be accessible before exiting.`)
 }
 
 func (f *flag) Validate() error {