+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added `wait` command to wait for various components of a test cluster to be ready.
+
 ### Fixed
 
 - Added logic for handling cluster creation errors that fail to create a cluster.
@@ -14,5 +18,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Modified `gsctl` execution to use the binary from the current `$PATH`.
+
+### Removed
+
+- Removed `--wait` flag from `create` command.
+- Removed unused `test` command.
 
 [Unreleased]: https://github.com/giantswarm/standup/tree/master